data_collect_yes

+++ rlinf/envs/realworld/common/wrappers/euler_obs.py
@@ -35,5 +35,5 @@ class Quat2EulerWrapper(gym.ObservationWrapper):
         tcp_pose = observation["state"]["tcp_pose"]
         observation["state"]["tcp_pose"] = np.concatenate(
             (tcp_pose[:3], R.from_quat(tcp_pose[3:].copy()).as_euler("xyz"))
-        )
+        ).astype(tcp_pose.dtype, copy=False)
         return observation
